PplPolicy refactoring
