+++ src/test/java/org/scion/jpan/ppl/PolicyTest.java
@@ -33,9 +33,9 @@ class PolicyTest {
 
     // "Empty policy"
     // Unlike the scionproto, we forbid empty policies.
-    assertThrows(PplException.class, () -> PplPolicy.builder().build());
+    assertThrows(PplException.class, () -> PplSubPolicy.builder().build());
     //    List<Path> paths = pp.getPaths("2-ff00:0:212", "2-ff00:0:211");
-    //    PplPolicy policy = PplPolicy.builder().build();
+    //    PplSubPolicy policy = PplSubPolicy.builder().build();
     //
     //    List<Path> outPaths = policy.filterAll(paths);
     //    assertEquals(2, outPaths.size());
@@ -44,12 +44,12 @@ class PolicyTest {
   @Test
   void testOptionsEval() {
     PathProvider pp = new PathProvider();
-    PplPolicy policy;
+    PplSubPolicy policy;
     List<Path> paths;
 
     // "one option, allow everything"
     policy =
-        PplPolicy.builder()
+        PplSubPolicy.builder()
             .addOption(
                 0,
                 PplExtPolicy.builder()
@@ -62,7 +62,7 @@ class PolicyTest {
 
     // "two options, deny everything"
     policy =
-        PplPolicy.builder()
+        PplSubPolicy.builder()
             .addOption(
                 0,
                 PplExtPolicy.builder()
@@ -81,7 +81,7 @@ class PolicyTest {
 
     // "two options, first: allow everything, second: allow one path"
     policy =
-        PplPolicy.builder()
+        PplSubPolicy.builder()
             .setName("")
             .addOption(
                 0,
@@ -103,7 +103,7 @@ class PolicyTest {
 
     // "two options, combined"
     policy =
-        PplPolicy.builder()
+        PplSubPolicy.builder()
             .addOption(
                 0,
                 PplExtPolicy.builder()
@@ -122,7 +122,7 @@ class PolicyTest {
 
     // "two options, take first"
     policy =
-        PplPolicy.builder()
+        PplSubPolicy.builder()
             .addOption(
                 1,
                 PplExtPolicy.builder()
@@ -141,7 +141,7 @@ class PolicyTest {
 
     // "two options, take second"
     policy =
-        PplPolicy.builder()
+        PplSubPolicy.builder()
             .addOption(
                 1,
                 PplExtPolicy.builder()
@@ -163,8 +163,8 @@ class PolicyTest {
   void TestExtends() {
     PplExtPolicy policy;
     PplExtPolicy[] extended;
-    PplPolicy extendedPolicy;
-    PplPolicy pol;
+    PplSubPolicy extendedPolicy;
+    PplSubPolicy pol;
 
     // "one extends, use sub acl"
     policy = PplExtPolicy.builder().addExtension("policy1").buildNoValidate();
@@ -177,9 +177,9 @@ class PolicyTest {
               .buildNoValidate()
         };
     extendedPolicy =
-        PplPolicy.builder().addAclEntry(true, "0-0#0").addAclEntry(DENY_STR).buildNoValidate();
+        PplSubPolicy.builder().addAclEntry(true, "0-0#0").addAclEntry(DENY_STR).buildNoValidate();
 
-    pol = PplPolicy.policyFromExtPolicy(policy, extended);
+    pol = PplSubPolicy.policyFromExtPolicy(policy, extended);
     assertEquals(extendedPolicy, pol);
 
     // "use option of extended policy"
@@ -197,7 +197,7 @@ class PolicyTest {
               .buildNoValidate()
         };
     extendedPolicy =
-        PplPolicy.builder()
+        PplSubPolicy.builder()
             .addOption(
                 1,
                 PplExtPolicy.builder()
@@ -205,7 +205,7 @@ class PolicyTest {
                     .addAclEntry(DENY_STR)
                     .buildNoValidate())
             .build();
-    pol = PplPolicy.policyFromExtPolicy(policy, extended);
+    pol = PplSubPolicy.policyFromExtPolicy(policy, extended);
     assertEquals(extendedPolicy, pol);
 
     // "two extends, use sub acl and list"
@@ -220,12 +220,12 @@ class PolicyTest {
               .buildNoValidate()
         };
     extendedPolicy =
-        PplPolicy.builder()
+        PplSubPolicy.builder()
             .addAclEntry(true, "0-0#0")
             .addAclEntry(DENY_STR)
             .setSequence("1-ff00:0:133#1019 1-ff00:0:132#1910")
             .buildNoValidate();
-    pol = PplPolicy.policyFromExtPolicy(policy, extended);
+    pol = PplSubPolicy.policyFromExtPolicy(policy, extended);
     assertEquals(extendedPolicy, pol);
 
     // "two extends, only use acl"
@@ -247,12 +247,12 @@ class PolicyTest {
               .buildNoValidate()
         };
     extendedPolicy =
-        PplPolicy.builder()
+        PplSubPolicy.builder()
             .addAclEntry(true, "0-0#0")
             .addAclEntry(DENY_STR)
             .setSequence("1-ff00:0:133#0 1-ff00:0:132#0")
             .buildNoValidate();
-    pol = PplPolicy.policyFromExtPolicy(policy, extended);
+    pol = PplSubPolicy.policyFromExtPolicy(policy, extended);
     assertEquals(extendedPolicy, pol);
 
     // "three extends, use last list"
@@ -272,8 +272,9 @@ class PolicyTest {
               .setSequence("1-ff00:0:133#1013 1-ff00:0:132#1913")
               .build()
         };
-    extendedPolicy = PplPolicy.builder().setSequence("1-ff00:0:133#1013 1-ff00:0:132#1913").build();
-    pol = PplPolicy.policyFromExtPolicy(policy, extended);
+    extendedPolicy =
+        PplSubPolicy.builder().setSequence("1-ff00:0:133#1013 1-ff00:0:132#1913").build();
+    pol = PplSubPolicy.policyFromExtPolicy(policy, extended);
     assertEquals(extendedPolicy, pol);
 
     // "nested extends"
@@ -287,8 +288,9 @@ class PolicyTest {
               .setSequence("1-ff00:0:133#1011 1-ff00:0:132#1911")
               .buildNoValidate()
         };
-    extendedPolicy = PplPolicy.builder().setSequence("1-ff00:0:133#1011 1-ff00:0:132#1911").build();
-    pol = PplPolicy.policyFromExtPolicy(policy, extended);
+    extendedPolicy =
+        PplSubPolicy.builder().setSequence("1-ff00:0:133#1011 1-ff00:0:132#1911").build();
+    pol = PplSubPolicy.policyFromExtPolicy(policy, extended);
     assertEquals(extendedPolicy, pol);
 
     // "nested extends, evaluating order"
@@ -306,8 +308,9 @@ class PolicyTest {
               .setSequence("1-ff00:0:133#1011 1-ff00:0:132#1911")
               .build()
         };
-    extendedPolicy = PplPolicy.builder().setSequence("1-ff00:0:133#1010 1-ff00:0:132#1910").build();
-    pol = PplPolicy.policyFromExtPolicy(policy, extended);
+    extendedPolicy =
+        PplSubPolicy.builder().setSequence("1-ff00:0:133#1010 1-ff00:0:132#1910").build();
+    pol = PplSubPolicy.policyFromExtPolicy(policy, extended);
     assertEquals(extendedPolicy, pol);
 
     // "different nested extends, evaluating order"
@@ -326,8 +329,9 @@ class PolicyTest {
               .setSequence("1-ff00:0:133#1011 1-ff00:0:132#1911")
               .build()
         };
-    extendedPolicy = PplPolicy.builder().setSequence("1-ff00:0:133#1010 1-ff00:0:132#1910").build();
-    pol = PplPolicy.policyFromExtPolicy(policy, extended);
+    extendedPolicy =
+        PplSubPolicy.builder().setSequence("1-ff00:0:133#1010 1-ff00:0:132#1910").build();
+    pol = PplSubPolicy.policyFromExtPolicy(policy, extended);
     assertEquals(extendedPolicy, pol);
 
     // "TestPolicy Extend not found"
@@ -341,7 +345,8 @@ class PolicyTest {
               .setSequence("1-ff00:0:133#1011 1-ff00:0:132#1911")
               .build()
         };
-    assertThrows(PplException.class, () -> PplPolicy.policyFromExtPolicy(errPolicy, errExtended));
+    assertThrows(
+        PplException.class, () -> PplSubPolicy.policyFromExtPolicy(errPolicy, errExtended));
   }
 
   @Test
@@ -351,13 +356,13 @@ class PolicyTest {
 
     // "sequence in options is ignored"
     PplExtPolicy optPolicy = PplExtPolicy.builder().setSequence("0+ 1-ff00:0:111 0+").build();
-    PplPolicy policy = PplPolicy.builder().addOption(0, optPolicy).build();
-    List<Path> outPaths = policy.filterOpt(paths, PplPolicy.FilterOptions.create(true));
+    PplSubPolicy policy = PplSubPolicy.builder().addOption(0, optPolicy).build();
+    List<Path> outPaths = policy.filterOpt(paths, PplSubPolicy.FilterOptions.create(true));
     assertEquals(3, outPaths.size());
 
     // "sequence is ignored"
-    policy = PplPolicy.builder().setSequence("0+ 1-ff00:0:111 0+").build();
-    outPaths = policy.filterOpt(paths, PplPolicy.FilterOptions.create(true));
+    policy = PplSubPolicy.builder().setSequence("0+ 1-ff00:0:111 0+").build();
+    outPaths = policy.filterOpt(paths, PplSubPolicy.FilterOptions.create(true));
     assertEquals(3, outPaths.size());
   }
